backend/src/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
aiosqlite
email-validator
redis
bcrypt==3.2.0
